[REFACTOR]: Resolved issues, Refresh Token-Access Token Issue, /me endpoint issue, added middleware to routes, removed one extra service file
