+++ src/routes/auth.routes.js
@@ -9,16 +9,16 @@ import {
   login,
   register,
 } from '../controllers/auth.controller.js';
+import { authenticateUser } from '../middlewares/authMiddleware.js';
 
 const router = express.Router();
 
 router.post('/forgot-password', forgotPasswordController);
 router.post('/reset-password/:token', resetPasswordController);
-router.post('/change-password', changePassword);
+router.post('/change-password', authenticateUser, changePassword);
 router.post('/refresh-token', refreshToken);
-router.post('/logout', logout);
-router.post('/me', getMe);
+router.get('/me', authenticateUser, getMe);
 router.post('/login', login);
 router.post('/register', register);
-
+router.post('/logout', authenticateUser, logout);
 export default router;
