+++ src/routes/index.js
@@ -1,5 +1,5 @@
 import { Router } from 'express';
-import AuthRouter from './auth.route.js';
+import AuthRouter from './auth.routes.js';
 const router = Router();
 router.use('/auth', AuthRouter);
