+++ src/services/auth.service.js
@@ -3,8 +3,7 @@ import jwt from 'jsonwebtoken';
 import mongoose from 'mongoose';
 import config from '../constants/config.js';
 import User from '../modals/User.js';
-import sendEmail from '../helpers/sendMail';
-
+import sendEmail from '../helpers/sendMail.js';
 export const forgotPasswordService = async email => {
   const user = await User.findOne({ email });
   if (!user) return;
@@ -74,20 +73,14 @@ export const changePasswordService = async (userId, oldPassword, newPassword) =>
 export const refreshTokenService = async refreshToken => {
   try {
     const decoded = jwt.verify(refreshToken, config.REFRESH_TOKEN_SECRET);
-    const user = await User.findById(decoded.id);
+    const user = await User.findById(decoded.userId);
     if (!user) {
       throw new Error('Invalid refresh token');
     }
 
-    const accessToken = jwt.sign({ id: user._id, email: user.email }, config.ACCESS_TOKEN_SECRET, {
-      expiresIn: config.ACCESS_TOKEN_EXPIRES_IN,
-    });
+    const accessToken = generateAccessToken(user);
 
-    const newRefreshToken = jwt.sign({ id: user._id }, config.REFRESH_TOKEN_SECRET, {
-      expiresIn: config.REFRESH_TOKEN_EXPIRES_IN,
-    });
-
-    return { accessToken, newRefreshToken };
+    return { accessToken };
   } catch (error) {
     if (error.name === 'JsonWebTokenError') {
       throw new Error('Invalid refresh token');
@@ -100,12 +93,8 @@ export const logoutService = async (userId, res) => {
   if (!mongoose.Types.ObjectId.isValid(userId)) {
     throw new Error('Invalid user ID');
   }
-
-  res.clearCookie('refreshToken', {
-    httpOnly: true,
-    sameSite: 'strict',
-  });
-
+  res.clearCookie('accessToken');
+  res.clearCookie('refreshToken');
   return true;
 };
 
@@ -120,4 +109,82 @@ export const getMeService = async userId => {
   }
 
   return user;
+};
+
+const generateAccessToken = user => {
+  return jwt.sign({ userId: user._id, role: user.role }, config.ACCESS_TOKEN_SECRET, {
+    expiresIn: config.ACCESS_TOKEN_EXPIRES_IN,
+  });
+};
+
+const generateRefreshToken = user => {
+  return jwt.sign({ userId: user._id, role: user.role }, config.REFRESH_TOKEN_SECRET, {
+    expiresIn: config.REFRESH_TOKEN_EXPIRES_IN,
+  });
+};
+
+export const loginUser = async (username, password) => {
+  if (!username || !password) {
+    const error = new Error('All fields are required');
+    error.statusCode = 400;
+    throw error;
+  }
+
+  const user = await User.findOne({
+    $or: [{ email: username }, { username: username }],
+  });
+
+  if (!user) {
+    const error = new Error('User not found');
+    error.statusCode = 401;
+    throw error;
+  }
+
+  const isPasswordValid = await bcrypt.compare(password, user.password);
+  if (!isPasswordValid) {
+    const error = new Error('Invalid credentials');
+    error.statusCode = 401;
+    throw error;
+  }
+
+  const accessToken = generateAccessToken(user);
+  const refreshToken = generateRefreshToken(user);
+
+  await User.findOneAndUpdate({ _id: user._id }, { refreshToken: refreshToken }, { new: true });
+
+  return {
+    accessToken,
+    refreshToken,
+  };
+};
+
+export const registerUser = async ({ username, email, password }) => {
+  if (!username || !email || !password) {
+    const error = new Error('All fields are required');
+    error.statusCode = 400;
+    throw error;
+  }
+
+  const existingUser = await User.findOne({ $or: [{ email }, { username }] });
+  if (existingUser) {
+    const error = new Error('Username or email already exists');
+    error.statusCode = 409;
+    throw error;
+  }
+
+  const hashedPassword = await bcrypt.hash(password, 10);
+
+  const newUser = new User({
+    username,
+    email,
+    password: hashedPassword,
+  });
+
+  await newUser.save();
+
+  return {
+    id: newUser._id,
+    username: newUser.username,
+    email: newUser.email,
+  };
 };
